add config in text format
Note that I've gone with duplication over DRY for this config for simplicity in debugging.
I'm sure that in the case of launching gvim, this could be handled more cleanly with a single
regex and more code, but I've settled for a version that works for all the described cases and is easy to debug.